deprecated dui from autocad
