+++ ConnectorAutocadCivil/ConnectorAutocadCivil.sln
@@ -11,8 +11,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Local Dependencies", "Local
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Core", "..\Core\Core\Core.csproj", "{3AD2C014-1ECA-4A6B-BAC4-30CE9A443A5B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DesktopUI", "..\DesktopUI\DesktopUI\DesktopUI.csproj", "{A97C3046-2D55-4C49-AB40-A3C2194B82FA}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Objects", "..\Objects\Objects\Objects.csproj", "{DE63379E-844A-47D9-AFDB-151A1888A912}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "ConverterAutocadCivil", "ConverterAutocadCivil", "{8AFDB1E6-D3E7-45EA-BCDC-F3554BDDEF1E}"
@@ -77,14 +75,6 @@ Global
 		{3AD2C014-1ECA-4A6B-BAC4-30CE9A443A5B}.Release|Any CPU.Build.0 = Release|Any CPU
 		{3AD2C014-1ECA-4A6B-BAC4-30CE9A443A5B}.Release|x64.ActiveCfg = Release|Any CPU
 		{3AD2C014-1ECA-4A6B-BAC4-30CE9A443A5B}.Release|x64.Build.0 = Release|Any CPU
-		{A97C3046-2D55-4C49-AB40-A3C2194B82FA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A97C3046-2D55-4C49-AB40-A3C2194B82FA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A97C3046-2D55-4C49-AB40-A3C2194B82FA}.Debug|x64.ActiveCfg = Debug|x64
-		{A97C3046-2D55-4C49-AB40-A3C2194B82FA}.Debug|x64.Build.0 = Debug|x64
-		{A97C3046-2D55-4C49-AB40-A3C2194B82FA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A97C3046-2D55-4C49-AB40-A3C2194B82FA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A97C3046-2D55-4C49-AB40-A3C2194B82FA}.Release|x64.ActiveCfg = Release|x64
-		{A97C3046-2D55-4C49-AB40-A3C2194B82FA}.Release|x64.Build.0 = Release|x64
 		{DE63379E-844A-47D9-AFDB-151A1888A912}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{DE63379E-844A-47D9-AFDB-151A1888A912}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DE63379E-844A-47D9-AFDB-151A1888A912}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -164,7 +154,6 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{58A88F1A-7489-46D2-949D-2FC3F68C8D84} = {A07071D5-E197-487D-B543-28639AC3C719}
 		{3AD2C014-1ECA-4A6B-BAC4-30CE9A443A5B} = {31CF26D2-2744-4A51-AA93-7F28408A006F}
-		{A97C3046-2D55-4C49-AB40-A3C2194B82FA} = {31CF26D2-2744-4A51-AA93-7F28408A006F}
 		{DE63379E-844A-47D9-AFDB-151A1888A912} = {31CF26D2-2744-4A51-AA93-7F28408A006F}
 		{8AFDB1E6-D3E7-45EA-BCDC-F3554BDDEF1E} = {31CF26D2-2744-4A51-AA93-7F28408A006F}
 		{8CFC7609-F640-4683-BF13-FE144D3DC50B} = {8AFDB1E6-D3E7-45EA-BCDC-F3554BDDEF1E}
