+++ ConnectorAutocadCivil/ConnectorAutocadCivil/ConnectorAutocadCivilShared.projitems
@@ -9,14 +9,11 @@
     <Import_RootNamespace>ConnectorAutoCADShared</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)Entry\SpeckleAutocadCommand2.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Storage\SpeckleStreamManager2.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)UI\ConnectorBindingsAutocadCivil2.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Utils.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Entry\App.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Entry\SpeckleAutocadCommand.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Storage\SpeckleStreamManager.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)UI\ConnectorBindingsAutocadCivil.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Utils.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Entry\App.cs" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="$(MSBuildThisFileDirectory)Resources\docs16.png" />
