+++ ConnectorAutocadCivil/ConnectorAutocadCivil/Entry/App.cs
@@ -40,13 +40,9 @@ namespace Speckle.ConnectorAutocadCivil.Entry
         //the below should fix it! This affects Avalonia and Material 
         AppDomain.CurrentDomain.AssemblyResolve += new ResolveEventHandler(OnAssemblyResolve);
 
-        // set up bindings and subscribe to document events
+        // DUI2
         SpeckleAutocadCommand.Bindings = new ConnectorBindingsAutocad();
-        SpeckleAutocadCommand.Bindings.SetExecutorAndInit();
-
-        // for DUI2
-        SpeckleAutocadCommand2.Bindings = new ConnectorBindingsAutocad2();
-        SpeckleAutocadCommand2.Bindings.RegisterAppEvents();
+        SpeckleAutocadCommand.Bindings.RegisterAppEvents();
       }
       catch(System.Exception e)
       {
@@ -96,10 +92,7 @@ namespace Speckle.ConnectorAutocadCivil.Entry
       if (panel == null)
         return;
       RibbonToolTip speckleTip = CreateToolTip("Speckle", "Speckle Connector for " + Utils.AppName);
-      RibbonButton button = CreateButton("Connector " + Utils.AppName + " (old)", "SpeckleOld", panel, null, speckleTip, "logo");
-
-      // DUI2
-      RibbonButton button2 = CreateButton("Connector " + Utils.AppName, "Speckle", panel, null, speckleTip, "logo");
+      RibbonButton button = CreateButton("Connector " + Utils.AppName, "Speckle", panel, null, speckleTip, "logo");
 
       // help and resources buttons
       RibbonSplitButton helpButton = new RibbonSplitButton();
@@ -174,8 +167,8 @@ namespace Speckle.ConnectorAutocadCivil.Entry
       button.ToolTip = tooltip;
       button.HelpSource = new System.Uri("https://speckle.guide/user/autocadcivil.html");
       button.Size = RibbonItemSize.Large;
-      button.Image = (CommandParameter == "SpeckleOld") ? LoadPngImgSource(imageName + "16_fade.png") : LoadPngImgSource(imageName + "16.png");
-      button.LargeImage = (CommandParameter == "SpeckleOld") ? LoadPngImgSource(imageName + "32_fade.png") : LoadPngImgSource(imageName + "32.png");
+      button.Image = LoadPngImgSource(imageName + "16.png");
+      button.LargeImage = LoadPngImgSource(imageName + "32.png");
 
       // add ribbon button pannel to the ribbon panel source
       if (sourcePanel != null)
@@ -229,20 +222,17 @@ namespace Speckle.ConnectorAutocadCivil.Entry
         if (btn != null)
           switch (commandParameter)
           {
-            case "SpeckleOld":
+            case "Speckle":
               SpeckleAutocadCommand.SpeckleCommand();
               break;
-            case "Speckle":
-              SpeckleAutocadCommand2.SpeckleCommand();
-              break;
             case "SpeckleCommunity":
-              SpeckleAutocadCommand2.SpeckleCommunity();
+              SpeckleAutocadCommand.SpeckleCommunity();
               break;
             case "SpeckleTutorials":
-              SpeckleAutocadCommand2.SpeckleTutorials();
+              SpeckleAutocadCommand.SpeckleTutorials();
               break;
             case "SpeckleDocs":
-              SpeckleAutocadCommand2.SpeckleDocs();
+              SpeckleAutocadCommand.SpeckleDocs();
               break;
           }
       }
