+++ ConnectorAutocadCivil/ConnectorAutocadCivil/Storage/SpeckleStreamManager.cs
@@ -1,64 +1,85 @@
 ﻿using System;
 using System.Collections.Generic;
-
+using System.Linq;
 using Autodesk.AutoCAD.ApplicationServices;
 using Autodesk.AutoCAD.DatabaseServices;
 
+using DesktopUI2.Models;
 using Speckle.Newtonsoft.Json;
 
 namespace Speckle.ConnectorAutocadCivil.Storage
 {
   /// <summary>
-  /// Manages streams stored in an autocad document
+  /// Manages the serialisation of speckle stream state
   /// </summary>
   /// <remarks>
   /// Uses a child dictionary for custom data in the Named Object Dictionary (NOD) which is the root level dictionary.
   /// This is because NOD persists after a document is closed (unlike file User Data).
   /// Custom data is stored as XRecord key value entries of type (string, ResultBuffer).
   /// ResultBuffers are TypedValue arrays, with the DxfCode of the input type as an integer.
+  /// Used for DesktopUI2
   /// </remarks>
   public static class SpeckleStreamManager
   {
-    readonly static string SpeckleExtensionDictionary = "SpeckleStreams";
+    readonly static string SpeckleExtensionDictionary = "Speckle";
+    readonly static string SpeckleStreamStates = "StreamStates";
 
-    public static List<string> GetSpeckleStreams()
+    // This is used to transfer old DUI saved streams to DUI2. Deprecate after a few releases
+    readonly static string SpeckleExtensionDictionaryOld = "SpeckleStreams";
+
+    /// <summary>
+    /// Returns all the speckle stream states present in the current document.
+    /// </summary>
+    /// <param name="doc"></param>
+    /// <returns></returns>
+    public static List<StreamState> ReadState(Document doc)
     {
-      Document Doc = Application.DocumentManager.MdiActiveDocument;
-      var streams = new List<string>();
+      var streams = new List<StreamState>();
 
-      if (Doc == null)
+      if (doc == null)
         return streams;
 
-      using (Transaction tr = Doc.Database.TransactionManager.StartTransaction())
+      using (Transaction tr = doc.Database.TransactionManager.StartTransaction())
       {
-        var NOD = (DBDictionary)tr.GetObject(Doc.Database.NamedObjectsDictionaryId, OpenMode.ForRead);
+        var NOD = (DBDictionary)tr.GetObject(doc.Database.NamedObjectsDictionaryId, OpenMode.ForRead);
         if (NOD.Contains(SpeckleExtensionDictionary))
         {
           var speckleDict = tr.GetObject(NOD.GetAt(SpeckleExtensionDictionary), OpenMode.ForRead) as DBDictionary;
           if (speckleDict != null && speckleDict.Count > 0)
           {
-            foreach (DBDictionaryEntry entry in speckleDict)
+            var id = speckleDict.GetAt(SpeckleStreamStates);
+            if (id != ObjectId.Null)
             {
-              var value = tr.GetObject(entry.Value, OpenMode.ForRead) as Xrecord;
-              streams.Add(value.Data.AsArray()[0].Value as string);
+              try // careful here: entries are length-capped and a serialized streamstate string could've been cut off, resulting in crash on deserialize
+              {
+                var record = tr.GetObject(id, OpenMode.ForRead) as Xrecord;
+                streams = JsonConvert.DeserializeObject<List<StreamState>>(record.Data.AsArray()[0].Value as string);
+              }
+              catch { }
             }
           }
         }
         tr.Commit();
       }
+
       return streams;
     }
 
-    public static void AddSpeckleStream(string id, string stream)
+    /// <summary>
+    /// Writes the stream states to the current document.
+    /// </summary>
+    /// <param name="doc"></param>
+    /// <param name="wrap"></param>
+    public static void WriteStreamStateList(Document doc, List<StreamState> streamStates)
     {
-      Document Doc = Application.DocumentManager.MdiActiveDocument;
-      if (Doc == null)
+      if (doc == null)
         return;
-      using (DocumentLock l = Doc.LockDocument())
+
+      using (DocumentLock l = doc.LockDocument())
       {
-        using (Transaction tr = Doc.Database.TransactionManager.StartTransaction())
+        using (Transaction tr = doc.Database.TransactionManager.StartTransaction())
         {
-          var NOD = (DBDictionary)tr.GetObject(Doc.Database.NamedObjectsDictionaryId, OpenMode.ForRead);
+          var NOD = (DBDictionary)tr.GetObject(doc.Database.NamedObjectsDictionaryId, OpenMode.ForRead);
           DBDictionary speckleDict;
           if (NOD.Contains(SpeckleExtensionDictionary))
           {
@@ -72,54 +93,51 @@ namespace Speckle.ConnectorAutocadCivil.Storage
             tr.AddNewlyCreatedDBObject(speckleDict, true);
           }
           var xRec = new Xrecord();
-          xRec.Data = new ResultBuffer(new TypedValue(Convert.ToInt32(DxfCode.Text), stream));
-          speckleDict.SetAt(id, xRec);
+          var value = JsonConvert.SerializeObject(streamStates) as string;
+          xRec.Data = new ResultBuffer(new TypedValue(Convert.ToInt32(DxfCode.Text), value));
+          speckleDict.SetAt(SpeckleStreamStates, xRec);
           tr.AddNewlyCreatedDBObject(xRec, true);
           tr.Commit();
         }
       }
     }
 
-    public static void RemoveSpeckleStream(string id)
+    // This is used to transfer old DUI saved streams to DUI2. Deprecate after a few releases
+    public static void TransferOldSpeckleStreams()
     {
       Document Doc = Application.DocumentManager.MdiActiveDocument;
-      if (Doc == null)
-        return;
-      using (DocumentLock l = Doc.LockDocument())
-      {
-        using (Transaction tr = Doc.Database.TransactionManager.StartTransaction())
-        {
-          var NOD = (DBDictionary)tr.GetObject(Doc.Database.NamedObjectsDictionaryId, OpenMode.ForRead);
-          if (NOD.Contains(SpeckleExtensionDictionary))
-          {
-            var speckleDict = (DBDictionary)tr.GetObject(NOD.GetAt(SpeckleExtensionDictionary), OpenMode.ForWrite);
-            speckleDict.Remove(id);
-          }
-          tr.Commit();
-        }
-      }
-    }
 
-    public static void UpdateSpeckleStream(string id, string stream)
-    {
-      Document Doc = Application.DocumentManager.MdiActiveDocument;
       if (Doc == null)
         return;
-      using (DocumentLock l = Doc.LockDocument())
+
+      List<string> oldSavedStreams = new List<string>();
+      using (Transaction tr = Doc.Database.TransactionManager.StartTransaction())
       {
-        using (Transaction tr = Doc.Database.TransactionManager.StartTransaction())
+        var NOD = (DBDictionary)tr.GetObject(Doc.Database.NamedObjectsDictionaryId, OpenMode.ForRead);
+        if (NOD.Contains(SpeckleExtensionDictionaryOld))
         {
-          var NOD = (DBDictionary)tr.GetObject(Doc.Database.NamedObjectsDictionaryId, OpenMode.ForRead);
-          if (NOD.Contains(SpeckleExtensionDictionary))
+          var speckleDict = tr.GetObject(NOD.GetAt(SpeckleExtensionDictionaryOld), OpenMode.ForWrite) as DBDictionary;
+          if (speckleDict != null && speckleDict.Count > 0)
           {
-            var speckleDict = (DBDictionary)tr.GetObject(NOD.GetAt(SpeckleExtensionDictionary), OpenMode.ForWrite);
-            var xRec = new Xrecord();
-            xRec.Data = new ResultBuffer(new TypedValue(Convert.ToInt32(DxfCode.Text), stream));
-            speckleDict.SetAt(id, xRec);
-            tr.AddNewlyCreatedDBObject(xRec, true);
+            // get old streams and add to list
+            foreach (DBDictionaryEntry entry in speckleDict)
+            {
+              var value = tr.GetObject(entry.Value, OpenMode.ForRead) as Xrecord;
+              oldSavedStreams.Add(value.Data.AsArray()[0].Value as string);
+            }
           }
-          tr.Commit();
+          NOD.Remove(SpeckleExtensionDictionaryOld); // removes old stream dictionary from doc
         }
+        tr.Commit();
+      }
+
+      if (oldSavedStreams.Count > 0)
+      {
+        // add old streams to new streamstate dictionary
+        var oldStreams = oldSavedStreams.Select(s => JsonConvert.DeserializeObject<StreamState>(s)).ToList();
+        var currentStreams = ReadState(Doc);
+        currentStreams.AddRange(oldStreams);
+        WriteStreamStateList(Doc, currentStreams);
       }
     }
   }
