+++ ConnectorAutocadCivil/ConnectorAutocadCivil/UI/ConnectorBindingsAutocadCivil.cs
@@ -1,96 +1,85 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Collections.Concurrent;
-using System.Threading.Tasks;
-using System.Text.RegularExpressions;
-using System.Collections;
-
-
-using Speckle.Newtonsoft.Json;
-using Speckle.Core.Models;
-using Speckle.Core.Kits;
-using Speckle.Core.Api;
-using Speckle.DesktopUI;
-using Speckle.DesktopUI.Utils;
-using ProgressReport = Speckle.DesktopUI.Utils.ProgressReport;
-using Speckle.Core.Transports;
+﻿using Autodesk.AutoCAD.ApplicationServices;
+using Autodesk.AutoCAD.Colors;
+using Autodesk.AutoCAD.DatabaseServices;
+using Autodesk.AutoCAD.EditorInput;
+using DesktopUI2;
+using DesktopUI2.Models;
+using DesktopUI2.Models.Filters;
+using DesktopUI2.Models.Settings;
+using DesktopUI2.ViewModels;
 using Speckle.ConnectorAutocadCivil.Entry;
 using Speckle.ConnectorAutocadCivil.Storage;
-
-using AcadApp = Autodesk.AutoCAD.ApplicationServices;
-using AcadDb = Autodesk.AutoCAD.DatabaseServices;
-using Autodesk.AutoCAD.EditorInput;
-using Autodesk.AutoCAD.ApplicationServices;
-using Autodesk.AutoCAD.Colors;
-
-using Stylet;
-using Autodesk.AutoCAD.DatabaseServices;
+using Speckle.Core.Api;
+using Speckle.Core.Kits;
+using Speckle.Core.Logging;
+using Speckle.Core.Models;
+using Speckle.Core.Transports;
+using System;
+using System.Collections;
+using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 
 namespace Speckle.ConnectorAutocadCivil.UI
 {
   public partial class ConnectorBindingsAutocad : ConnectorBindings
   {
-    public Document Doc => Application.DocumentManager.MdiActiveDocument;
-
-    /// <summary>
-    /// TODO: Any errors thrown should be stored here and passed to the ui state
-    /// </summary>
-    public List<Exception> Exceptions { get; set; } = new List<Exception>();
+    public static Document Doc => Application.DocumentManager.MdiActiveDocument;
 
     // AutoCAD API should only be called on the main thread.
     // Not doing so results in botched conversions for any that require adding objects to Document model space before modifying (eg adding vertices and faces for meshes)
     // There's no easy way to access main thread from document object, therefore we are creating a control during Connector Bindings constructor (since it's called on main thread) that allows for invoking worker threads on the main thread
-    public System.Windows.Forms.Control Control; 
+    public System.Windows.Forms.Control Control;
+
     public ConnectorBindingsAutocad() : base()
     {
       Control = new System.Windows.Forms.Control();
       Control.CreateControl();
     }
 
-    public void SetExecutorAndInit()
-    {
-      Application.DocumentManager.DocumentActivated += Application_DocumentActivated;
-      Doc.BeginDocumentClose += Application_DocumentClosed;
-    }
-
     #region local streams 
-
-    public override void AddNewStream(StreamState state)
+    public override void WriteStreamsToFile(List<StreamState> streams)
     {
-      SpeckleStreamManager.AddSpeckleStream(state.Stream.id, JsonConvert.SerializeObject(state));
-    }
-
-    public override void RemoveStreamFromFile(string streamId)
-    {
-      SpeckleStreamManager.RemoveSpeckleStream(streamId);
-    }
-
-    public override void PersistAndUpdateStreamInFile(StreamState state)
-    {
-      SpeckleStreamManager.UpdateSpeckleStream(state.Stream.id, JsonConvert.SerializeObject(state));
+      SpeckleStreamManager.WriteStreamStateList(Doc, streams);
     }
 
     public override List<StreamState> GetStreamsInFile()
     {
-      List<string> strings = SpeckleStreamManager.GetSpeckleStreams();
-      return strings.Select(s => JsonConvert.DeserializeObject<StreamState>(s)).ToList();
+      var streams = new List<StreamState>();
+      if (Doc != null)
+        streams = SpeckleStreamManager.ReadState(Doc);
+      return streams;
     }
     #endregion
 
     #region boilerplate
+    public override string GetHostAppNameVersion() => Utils.VersionedAppName.Replace("AutoCAD", "AutoCAD ").Replace("Civil", "Civil 3D  "); //hack for ADSK store;
+    public override string GetHostAppName() => Utils.Slug;
 
-    public override string GetActiveViewName()
+
+
+    private string GetDocPath(Document doc) => HostApplicationServices.Current.FindFile(doc?.Name, doc?.Database, FindFileHint.Default);
+
+    public override string GetDocumentId()
     {
-      return "Entire Document"; // TODO: handle views
+      string path = GetDocPath(Doc);
+      var hash = Core.Models.Utilities.hashString(path + Doc?.Name, Core.Models.Utilities.HashingFuctions.MD5);
+      return hash;
     }
 
-    public override List<string> GetObjectsInView() // TODO: this returns all visible doc objects. handle views later.
+    public override string GetDocumentLocation() => GetDocPath(Doc);
+
+    public override string GetFileName() => Doc?.Name;
+
+    public override string GetActiveViewName() => "Entire Document";
+
+    public override List<string> GetObjectsInView() // this returns all visible doc objects.
     {
       var objs = new List<string>();
       using (Transaction tr = Doc.Database.TransactionManager.StartTransaction())
       {
-        BlockTableRecord modelSpace = Doc.Database.GetModelSpace(); 
+        BlockTableRecord modelSpace = Doc.Database.GetModelSpace();
         foreach (ObjectId id in modelSpace)
         {
           var dbObj = tr.GetObject(id, OpenMode.ForRead);
@@ -101,18 +90,6 @@ namespace Speckle.ConnectorAutocadCivil.UI
       }
       return objs;
     }
-
-    public override string GetHostAppName() => Utils.VersionedAppName.Replace("AutoCAD", "AutoCAD ").Replace("Civil", "Civil 3D  "); //hack for ADSK store;
-
-    public override string GetDocumentId()
-    {
-      string path = HostApplicationServices.Current.FindFile(Doc.Name, Doc.Database, FindFileHint.Default);
-      return Core.Models.Utilities.hashString("X" + path + Doc?.Name, Core.Models.Utilities.HashingFuctions.MD5); // what is the "X" prefix for?
-    }
-
-    public override string GetDocumentLocation() => AcadDb.HostApplicationServices.Current.FindFile(Doc.Name, Doc.Database, AcadDb.FindFileHint.Default);
-
-    public override string GetFileName() => Doc?.Name;
 
     public override List<string> GetSelectedObjects()
     {
@@ -144,9 +121,21 @@ namespace Speckle.ConnectorAutocadCivil.UI
       }
       return new List<ISelectionFilter>()
       {
-         new ListSelectionFilter {Slug="layer",  Name = "Layers", Icon = "LayersTriple", Description = "Selects objects based on their layers.", Values = layers },
-         new AllSelectionFilter {Slug="all",  Name = "All", Icon = "CubeScan", Description = "Selects all document objects." }
+        new ManualSelectionFilter(),
+        new ListSelectionFilter {Slug="layer",  Name = "Layers", Icon = "LayersTriple", Description = "Selects objects based on their layers.", Values = layers },
+        new AllSelectionFilter {Slug="all",  Name = "All", Icon = "CubeScan", Description = "Selects all document objects." }
       };
+    }
+
+    public override List<ISetting> GetSettings()
+    {
+      return new List<ISetting>();
+    }
+
+    //TODO
+    public override List<MenuItem> GetCustomStreamMenuItems()
+    {
+      return new List<MenuItem>();
     }
 
     public override void SelectClientObjects(string args)
@@ -154,100 +143,83 @@ namespace Speckle.ConnectorAutocadCivil.UI
       throw new NotImplementedException();
     }
 
-    private void UpdateProgress(ConcurrentDictionary<string, int> dict, ProgressReport progress)
-    {
-      if (progress == null)
-      {
-        return;
-      }
-
-      Execute.PostToUIThread(() =>
-      {
-        progress.ProgressDict = dict;
-        progress.Value = dict.Values.Last();
-      });
-    }
-
     #endregion
 
     #region receiving 
-
-    public override async Task<StreamState> ReceiveStream(StreamState state)
+    public override async Task<StreamState> ReceiveStream(StreamState state, ProgressViewModel progress)
     {
-      if (Doc == null)
-      {
-        state.Errors.Add(new Exception($"No Document is open."));
-        return null;
-      }
-
-      Exceptions.Clear();
-
       var kit = KitManager.GetDefaultKit();
       var converter = kit.LoadConverter(Utils.VersionedAppName);
-      var transport = new ServerTransport(state.Client.Account, state.Stream.id);
+      if (converter == null)
+        throw new Exception("Could not find any Kit!");
+      var transport = new ServerTransport(state.Client.Account, state.StreamId);
 
-      var stream = await state.Client.StreamGet(state.Stream.id);
+      var stream = await state.Client.StreamGet(state.StreamId);
 
-      if (state.CancellationTokenSource.Token.IsCancellationRequested)
-      {
+      if (progress.CancellationTokenSource.Token.IsCancellationRequested)
         return null;
+
+      if (Doc == null)
+      {
+        progress.Report.LogOperationError(new Exception($"No Document is open."));
+        progress.CancellationTokenSource.Cancel();
       }
-
-      string referencedObject = state.Commit.referencedObject;
-
-      var commitId = state.Commit.id;
-      var commitMsg = state.Commit.message;
 
       //if "latest", always make sure we get the latest commit when the user clicks "receive"
-      if (commitId == "latest")
+      Commit commit = null;
+      if (state.CommitId == "latest")
       {
-        var res = await state.Client.BranchGet(state.CancellationTokenSource.Token, state.Stream.id, state.Branch.name, 1);
-        var commit = res.commits.items.FirstOrDefault();
-        commitId = commit.id;
-        commitMsg = commit.message;
-        referencedObject = commit.referencedObject;
+        var res = await state.Client.BranchGet(progress.CancellationTokenSource.Token, state.StreamId, state.BranchName, 1);
+        commit = res.commits.items.FirstOrDefault();
       }
-
-      var commitObject = await Operations.Receive(
-        referencedObject,
-        state.CancellationTokenSource.Token,
-        transport,
-        onProgressAction: d => UpdateProgress(d, state.Progress),
-        onTotalChildrenCountKnown: num => Execute.PostToUIThread(() => state.Progress.Maximum = num),
-        onErrorAction: (message, exception) => { Exceptions.Add(exception); },
-        disposeTransports: true
-        );
-      
+      else
+      {
+        commit = await state.Client.CommitGet(progress.CancellationTokenSource.Token, state.StreamId, state.CommitId);
+      }
+      string referencedObject = commit.referencedObject;
+      Base commitObject = null;
       try
       {
+        commitObject = await Operations.Receive(
+          referencedObject,
+          progress.CancellationTokenSource.Token,
+          transport,
+          onProgressAction: dict => progress.Update(dict),
+          onErrorAction: (s, e) =>
+          {
+            progress.Report.LogOperationError(e);
+            progress.CancellationTokenSource.Cancel();
+          },
+          onTotalChildrenCountKnown: count => { progress.Max = count; },
+          disposeTransports: true
+          );
+
         await state.Client.CommitReceived(new CommitReceivedInput
         {
-          streamId = stream.id,
-          commitId = commitId,
-          message = commitMsg,
+          streamId = stream?.id,
+          commitId = commit?.id,
+          message = commit?.message,
           sourceApplication = Utils.VersionedAppName
         });
       }
-      catch
+      catch (Exception e)
       {
-        // Do nothing!
+        progress.Report.OperationErrors.Add(new Exception($"Could not receive or deserialize commit: {e.Message}"));
       }
-      if (Exceptions.Count != 0)
-      {
-        RaiseNotification($"Encountered error: {Exceptions.Last().Message}");
-      }
+      if (progress.Report.OperationErrorsCount != 0 || commitObject == null)
+        return state;
 
       // invoke conversions on the main thread via control
       if (Control.InvokeRequired)
-        Control.Invoke(new ConversionDelegate(ConvertCommit), new object[] { commitObject, converter, state, stream, commitId });
+        Control.Invoke(new ReceivingDelegate(ConvertReceiveCommit), new object[] { commitObject, converter, state, progress, stream, commit.id });
       else
-        ConvertCommit(commitObject, converter, state, stream, commitId);
+        ConvertReceiveCommit(commitObject, converter, state, progress, stream, commit.id);
 
       return state;
     }
 
-    delegate void ConversionDelegate(Base commitObject, ISpeckleConverter converter, StreamState state, Stream stream, string id);
-    private void ConvertCommit(Base commitObject, ISpeckleConverter converter, StreamState state, Stream stream, string id)
+    delegate void ReceivingDelegate(Base commitObject, ISpeckleConverter converter, StreamState state, ProgressViewModel progress, Stream stream, string id);
+    private void ConvertReceiveCommit(Base commitObject, ISpeckleConverter converter, StreamState state, ProgressViewModel progress, Stream stream, string id)
     {
       using (DocumentLock l = Doc.LockDocument())
       {
@@ -258,19 +230,13 @@ namespace Speckle.ConnectorAutocadCivil.UI
 
           // keep track of conversion progress here
           var conversionProgressDict = new ConcurrentDictionary<string, int>();
-          conversionProgressDict["Conversion"] = 0;
-          Execute.PostToUIThread(() => state.Progress.Maximum = state.SelectedObjectIds.Count());
-          Action updateProgressAction = () =>
-          {
-            conversionProgressDict["Conversion"]++;
-            UpdateProgress(conversionProgressDict, state.Progress);
-          };
+          conversionProgressDict["Conversion"] = 1;
 
           // keep track of any layer name changes for notification here
           bool changedLayerNames = false;
 
-          // create a commit layer prefix: all nested layers will be concatenated with this
-          var commitPrefix = DesktopUI.Utils.Formatting.CommitInfo(stream.name, state.Branch.name, id);
+          // create a commit prefix: used for layers and block definition names
+          var commitPrefix = Formatting.CommitInfo(stream.name, state.BranchName, id);
 
           // give converter a way to access the commit info
           if (Doc.UserData.ContainsKey("commit"))
@@ -278,7 +244,7 @@ namespace Speckle.ConnectorAutocadCivil.UI
           else
             Doc.UserData.Add("commit", commitPrefix);
 
-          // delete existing commit layers and block instances
+          // delete existing commit layers
           try
           {
             DeleteBlocksWithPrefix(commitPrefix, tr);
@@ -286,15 +252,16 @@ namespace Speckle.ConnectorAutocadCivil.UI
           }
           catch
           {
-            RaiseNotification($"could not remove existing layers or blocks starting with {commitPrefix} before importing new geometry.");
-            state.Errors.Add(new Exception($"could not remove existing layers or blocks starting with {commitPrefix} before importing new geometry."));
+            converter.Report.LogOperationError(new Exception($"Failed to remove existing layers or blocks starting with {commitPrefix} before importing new geometry."));
           }
 
           // flatten the commit object to retrieve children objs
           int count = 0;
           var commitObjs = FlattenCommitObject(commitObject, converter, commitPrefix, state, ref count);
 
-          // TODO: create dictionaries here for linetype and layer linewidth
+          // open model space block table record for write
+          BlockTableRecord btr = (BlockTableRecord)tr.GetObject(Doc.Database.CurrentSpaceId, OpenMode.ForWrite);
+
           // More efficient this way than doing this per object
           var lineTypeDictionary = new Dictionary<string, ObjectId>();
           var lineTypeTable = (LinetypeTable)tr.GetObject(Doc.Database.LinetypeTableId, OpenMode.ForRead);
@@ -309,7 +276,19 @@ namespace Speckle.ConnectorAutocadCivil.UI
             // create the object's bake layer if it doesn't already exist
             (Base obj, string layerName) = commitObj;
 
-            var converted = converter.ConvertToNative(obj);
+            conversionProgressDict["Conversion"]++;
+            progress.Update(conversionProgressDict);
+
+            object converted = null;
+            try
+            {
+              converted = converter.ConvertToNative(obj);
+            }
+            catch (Exception e)
+            {
+              progress.Report.LogConversionError(new Exception($"Failed to convert object {obj.id} of type {obj.speckle_type}: {e.Message}"));
+              continue;
+            }
             var convertedEntity = converted as Entity;
 
             if (convertedEntity != null)
@@ -320,56 +299,34 @@ namespace Speckle.ConnectorAutocadCivil.UI
                 if (!cleanName.Equals(layerName))
                   changedLayerNames = true;
 
-                var appended = convertedEntity.Append(cleanName);
-                if (appended.IsValid)
+                var res = convertedEntity.Append(cleanName);
+                if (res.IsValid)
                 {
-                  // handle display
+                  // handle display - fallback to rendermaterial if no displaystyle exists
                   Base display = obj[@"displayStyle"] as Base;
-                  if (display != null)
-                  {
-                    var color = display["color"] as int?;
-                    var lineType = display["linetype"] as string;
-                    var lineWidth = display["lineweight"] as double?;
+                  if (display == null) display = obj[@"renderMaterial"] as Base;
+                  if (display != null) Utils.SetStyle(display, convertedEntity, lineTypeDictionary);
 
-                    if (color != null)
-                    {
-                      var systemColor = System.Drawing.Color.FromArgb((int)color);
-                      convertedEntity.Color = Color.FromRgb(systemColor.R, systemColor.G, systemColor.B);
-                      convertedEntity.Transparency = new Transparency(systemColor.A);
-                    }
-                    if (lineWidth != null)
-                    {
-                      convertedEntity.LineWeight = Utils.GetLineWeight((double)lineWidth);
-                    }
-
-                    if (lineType != null)
-                    {
-                      if (lineTypeDictionary.ContainsKey(lineType))
-                      {
-                        convertedEntity.LinetypeId = lineTypeDictionary[lineType];
-                      }
-                    }
-                  }
                   tr.TransactionManager.QueueForGraphicsFlush();
                 }
                 else
                 {
-                  state.Errors.Add(new Exception($"Failed to bake object {obj.id} of type {obj.speckle_type}."));
+                  progress.Report.LogConversionError(new Exception($"Failed to add converted object {obj.id} of type {obj.speckle_type} to the document."));
                 }
 
               }
               else
-                state.Errors.Add(new Exception($"Could not create layer {layerName} to bake objects into."));
+                progress.Report.LogOperationError(new Exception($"Failed to create layer {layerName} to bake objects into."));
             }
             else if (converted == null)
             {
-              state.Errors.Add(new Exception($"Failed to convert object {obj.id} of type {obj.speckle_type}."));
+              progress.Report.LogConversionError(new Exception($"Failed to convert object {obj.id} of type {obj.speckle_type}."));
             }
           }
+          progress.Report.Merge(converter.Report);
 
-          // raise any warnings from layer name modification
           if (changedLayerNames)
-            state.Errors.Add(new Exception($"Layer names were modified: one or more layers contained invalid characters {Utils.invalidChars}"));
+            progress.Report.Log($"Layer names were modified: one or more layers contained invalid characters {Utils.invalidChars}");
 
           // remove commit info from doc userdata
           Doc.UserData.Remove("commit");
@@ -417,7 +374,7 @@ namespace Speckle.ConnectorAutocadCivil.UI
             }
           }
           if (!foundConvertibleMember && count == totalMembers) // this was an unsupported geo
-            state.Errors.Add(new Exception($"Receiving {@base.speckle_type} objects is not supported. Object {@base.id} not baked."));
+            converter.Report.Log($"Skipped not supported type: { @base.speckle_type }. Object {@base.id} not baked.");
           return objects;
         }
       }
@@ -441,13 +398,27 @@ namespace Speckle.ConnectorAutocadCivil.UI
       return objects;
     }
 
-    private void DeleteLayersWithPrefix(string prefix, AcadDb.Transaction tr)
+    private void DeleteBlocksWithPrefix(string prefix, Transaction tr)
+    {
+      BlockTable blockTable = tr.GetObject(Doc.Database.BlockTableId, OpenMode.ForRead) as BlockTable;
+      foreach (ObjectId blockId in blockTable)
+      {
+        BlockTableRecord block = (BlockTableRecord)tr.GetObject(blockId, OpenMode.ForRead);
+        if (block.Name.StartsWith(prefix))
+        {
+          block.UpgradeOpen();
+          block.Erase();
+        }
+      }
+    }
+
+    private void DeleteLayersWithPrefix(string prefix, Transaction tr)
     {
       // Open the Layer table for read
-      var lyrTbl = (AcadDb.LayerTable)tr.GetObject(Doc.Database.LayerTableId, AcadDb.OpenMode.ForRead);
-      foreach (AcadDb.ObjectId layerId in lyrTbl)
+      var lyrTbl = (LayerTable)tr.GetObject(Doc.Database.LayerTableId, OpenMode.ForRead);
+      foreach (ObjectId layerId in lyrTbl)
       {
-        AcadDb.LayerTableRecord layer = (AcadDb.LayerTableRecord)tr.GetObject(layerId, AcadDb.OpenMode.ForRead);
+        LayerTableRecord layer = (LayerTableRecord)tr.GetObject(layerId, OpenMode.ForRead);
         string layerName = layer.Name;
         if (layerName.StartsWith(prefix))
         {
@@ -463,13 +434,13 @@ namespace Speckle.ConnectorAutocadCivil.UI
 
           // delete all objects on this layer
           // TODO: this is ugly! is there a better way to delete layer objs instead of looping through each one?
-          var bt = (AcadDb.BlockTable)tr.GetObject(Doc.Database.BlockTableId, AcadDb.OpenMode.ForRead);
+          var bt = (BlockTable)tr.GetObject(Doc.Database.BlockTableId, OpenMode.ForRead);
           foreach (var btId in bt)
           {
-            var block = (AcadDb.BlockTableRecord)tr.GetObject(btId, AcadDb.OpenMode.ForRead);
+            var block = (BlockTableRecord)tr.GetObject(btId, OpenMode.ForRead);
             foreach (var entId in block)
             {
-              var ent = (AcadDb.Entity)tr.GetObject(entId, AcadDb.OpenMode.ForRead);
+              var ent = (Entity)tr.GetObject(entId, OpenMode.ForRead);
               if (ent.Layer == layerName)
               {
                 ent.UpgradeOpen();
@@ -482,41 +453,28 @@ namespace Speckle.ConnectorAutocadCivil.UI
         }
       }
     }
-    private void DeleteBlocksWithPrefix(string prefix, Transaction tr)
-    {
-      BlockTable blockTable = tr.GetObject(Doc.Database.BlockTableId, OpenMode.ForRead) as BlockTable;
-      foreach (ObjectId blockId in blockTable)
-      {
-        BlockTableRecord block = (BlockTableRecord)tr.GetObject(blockId, OpenMode.ForRead);
-        if (block.Name.StartsWith(prefix))
-        {
-          block.UpgradeOpen();
-          block.Erase();
-        }
-      }
-    }
 
-    private bool GetOrMakeLayer(string layerName, AcadDb.Transaction tr, out string cleanName)
+    private bool GetOrMakeLayer(string layerName, Transaction tr, out string cleanName)
     {
       cleanName = Utils.RemoveInvalidChars(layerName);
       try
       {
-        AcadDb.LayerTable lyrTbl = tr.GetObject(Doc.Database.LayerTableId, AcadDb.OpenMode.ForRead) as AcadDb.LayerTable;
-        if (lyrTbl.Has(cleanName))
+        LayerTable layerTable = tr.GetObject(Doc.Database.LayerTableId, OpenMode.ForRead) as LayerTable;
+        if (layerTable.Has(cleanName))
         {
           return true;
         }
         else
         {
-          lyrTbl.UpgradeOpen();
-          var _layer = new AcadDb.LayerTableRecord();
+          layerTable.UpgradeOpen();
+          var _layer = new LayerTableRecord();
 
           // Assign the layer properties
           _layer.Color = Autodesk.AutoCAD.Colors.Color.FromColorIndex(ColorMethod.ByColor, 7); // white
           _layer.Name = cleanName;
 
           // Append the new layer to the layer table and the transaction
-          lyrTbl.Add(_layer);
+          layerTable.Add(_layer);
           tr.AddNewlyCreatedDBObject(_layer, true);
         }
       }
@@ -527,99 +485,155 @@ namespace Speckle.ConnectorAutocadCivil.UI
     #endregion
 
     #region sending
-
-    public override async Task<StreamState> SendStream(StreamState state)
+    public override async Task<string> SendStream(StreamState state, ProgressViewModel progress)
     {
       var kit = KitManager.GetDefaultKit();
       var converter = kit.LoadConverter(Utils.VersionedAppName);
-      
-      var streamId = state.Stream.id;
+      var streamId = state.StreamId;
       var client = state.Client;
 
       if (state.Filter != null)
-      {
         state.SelectedObjectIds = GetObjectsFromFilter(state.Filter, converter);
-      }
 
       // remove deleted object ids
       var deletedElements = new List<string>();
       foreach (var handle in state.SelectedObjectIds)
-        if (Doc.Database.TryGetObjectId(Utils.GetHandle(handle), out AcadDb.ObjectId id))
+        if (Doc.Database.TryGetObjectId(Utils.GetHandle(handle), out ObjectId id))
           if (id.IsErased || id.IsNull)
             deletedElements.Add(handle);
       state.SelectedObjectIds = state.SelectedObjectIds.Where(o => !deletedElements.Contains(o)).ToList();
 
       if (state.SelectedObjectIds.Count == 0)
       {
-        RaiseNotification("Zero objects selected; send stopped. Please select some objects, or check that your filter can actually select something.");
-        return state;
+        progress.Report.LogOperationError(new Exception("Zero objects selected; send stopped. Please select some objects, or check that your filter can actually select something."));
+        return null;
       }
 
-      var commitObj = new Base();
+      var commitObject = new Base();
+      commitObject["units"] = Utils.GetUnits(Doc); // TODO: check whether commits base needs units attached
 
-      /* Deprecated until we decide whether or not commit objs need units. If so, should add UnitsToSpeckle conversion method to connector
-      var units = Units.GetUnitsFromString(Doc.Database.Insunits.ToString());
-      commitObj["units"] = units;
-      */
-
-      var conversionProgressDict = new ConcurrentDictionary<string, int>();
-      conversionProgressDict["Conversion"] = 0;
-      Execute.PostToUIThread(() => state.Progress.Maximum = state.SelectedObjectIds.Count());
       int convertedCount = 0;
-      bool renamedlayers = false;
 
+      // invoke conversions on the main thread via control
+      if (Control.InvokeRequired)
+        Control.Invoke(new Action(() => ConvertSendCommit(commitObject, converter, state, progress, ref convertedCount)), new object[] { });
+      else
+        ConvertSendCommit(commitObject, converter, state, progress, ref convertedCount);
+
+      progress.Report.Merge(converter.Report);
+
+      if (convertedCount == 0)
+      {
+        progress.Report.LogOperationError(new SpeckleException("Zero objects converted successfully. Send stopped.", false));
+        return null;
+      }
+
+      if (progress.CancellationTokenSource.Token.IsCancellationRequested)
+        return null;
+
+      var transports = new List<ITransport>() { new ServerTransport(client.Account, streamId) };
+
+      var commitObjId = await Operations.Send(
+        commitObject,
+        progress.CancellationTokenSource.Token,
+        transports,
+        onProgressAction: dict => progress.Update(dict),
+        onErrorAction: (err, exception) =>
+        {
+          progress.Report.LogOperationError(exception);
+          progress.CancellationTokenSource.Cancel();
+        },
+        disposeTransports: true
+        );
+
+      if (progress.Report.OperationErrorsCount != 0)
+        return null;
+
+      var actualCommit = new CommitCreateInput
+      {
+        streamId = streamId,
+        objectId = commitObjId,
+        branchName = state.BranchName,
+        message = state.CommitMessage != null ? state.CommitMessage : $"Pushed {convertedCount} elements from {Utils.AppName}.",
+        sourceApplication = Utils.VersionedAppName
+      };
+
+      if (state.PreviousCommitId != null) { actualCommit.parents = new List<string>() { state.PreviousCommitId }; }
+
+      try
+      {
+        var commitId = await client.CommitCreate(actualCommit);
+        state.PreviousCommitId = commitId;
+        return commitId;
+      }
+      catch (Exception e)
+      {
+        progress.Report.LogOperationError(e);
+      }
+      return null;
+    }
+
+    delegate void SendingDelegate(Base commitObject, ISpeckleConverter converter, StreamState state, ProgressViewModel progress, ref int convertedCount);
+    private void ConvertSendCommit(Base commitObject, ISpeckleConverter converter, StreamState state, ProgressViewModel progress, ref int convertedCount)
+    {
       using (Transaction tr = Doc.Database.TransactionManager.StartTransaction())
       {
-        converter.SetContextDocument(Doc); // set context doc here to capture transaction prop
+        // set the context doc for conversion - this is set inside the transaction loop because the converter retrieves this transaction for all db editing when the context doc is set!
+        converter.SetContextDocument(Doc);
+
+        var conversionProgressDict = new ConcurrentDictionary<string, int>();
+        conversionProgressDict["Conversion"] = 0;
+
+        bool renamedlayers = false;
 
         foreach (var autocadObjectHandle in state.SelectedObjectIds)
         {
-          if (state.CancellationTokenSource.Token.IsCancellationRequested)
+          if (progress.CancellationTokenSource.Token.IsCancellationRequested)
           {
-            return null;
+            tr.Commit();
+            return;
           }
 
+          conversionProgressDict["Conversion"]++;
+          progress.Update(conversionProgressDict);
+
           // get the db object from id
-          AcadDb.Handle hn = Utils.GetHandle(autocadObjectHandle);
-          AcadDb.DBObject obj = hn.GetObject(tr, out string type, out string layer);
+          Handle hn = Utils.GetHandle(autocadObjectHandle);
+          DBObject obj = hn.GetObject(tr, out string type, out string layer);
 
           if (obj == null)
           {
-            state.Errors.Add(new Exception($"Failed to find local object ${autocadObjectHandle}."));
+            progress.Report.Log($"Skipped not found object: ${autocadObjectHandle}.");
             continue;
           }
 
           if (!converter.CanConvertToSpeckle(obj))
           {
-            state.Errors.Add(new Exception($"Objects of type ${type} are not supported"));
+            progress.Report.Log($"Skipped not supported type: ${type}. Object ${obj.Id} not sent.");
             continue;
           }
 
           // convert obj
-          // try catch to prevent memory access violation crash in case a conversion goes wrong
           Base converted = null;
           string containerName = string.Empty;
-          try
+          converted = converter.ConvertToSpeckle(obj);
+          if (converted == null)
           {
-            converted = converter.ConvertToSpeckle(obj);
-            if (converted == null)
-            {
-              state.Errors.Add(new Exception($"Failed to convert object ${autocadObjectHandle} of type ${type}."));
-              continue;
-            }
-          }
-          catch
-          {
-            state.Errors.Add(new Exception($"Failed to convert object {autocadObjectHandle} of type {type}."));
+            progress.Report.LogConversionError(new Exception($"Failed to convert object {autocadObjectHandle} of type {type}."));
             continue;
           }
 
           /* TODO: adding the extension dictionary / xdata per object 
           foreach (var key in obj.ExtensionDictionary)
-          {
             converted[key] = obj.ExtensionDictionary.GetUserString(key);
-          }
           */
+
+#if CIVIL2021 || CIVIL2022
+          // add property sets if this is Civil3D
+          var propertySets = obj.GetPropertySets(tr);
+          if (propertySets.Count > 0)
+            converted["propertySets"] = propertySets;
+#endif
 
           if (obj is BlockReference)
             containerName = "Blocks";
@@ -632,110 +646,66 @@ namespace Speckle.ConnectorAutocadCivil.UI
               renamedlayers = true;
           }
 
-          if (commitObj[$"@{containerName}"] == null)
-            commitObj[$"@{containerName}"] = new List<Base>();
-          ((List<Base>)commitObj[$"@{containerName}"]).Add(converted);
+          if (commitObject[$"@{containerName}"] == null)
+            commitObject[$"@{containerName}"] = new List<Base>();
+          ((List<Base>)commitObject[$"@{containerName}"]).Add(converted);
 
           conversionProgressDict["Conversion"]++;
-          UpdateProgress(conversionProgressDict, state.Progress);
+          progress.Update(conversionProgressDict);
 
           converted.applicationId = autocadObjectHandle;
 
           convertedCount++;
         }
 
+        if (renamedlayers)
+          progress.Report.Log("Replaced illegal chars ./ with - in one or more layer names.");
+
         tr.Commit();
       }
-
-      if (renamedlayers)
-        RaiseNotification("Replaced illegal chars ./ with - in one or more layer names.");
-
-      if (state.CancellationTokenSource.Token.IsCancellationRequested)
-      {
-        return null;
-      }
-
-      Execute.PostToUIThread(() => state.Progress.Maximum = convertedCount);
-
-      var transports = new List<ITransport>() { new ServerTransport(client.Account, streamId) };
-
-      var commitObjId = await Operations.Send(
-        commitObj,
-        state.CancellationTokenSource.Token,
-        transports,
-        onProgressAction: dict => UpdateProgress(dict, state.Progress),
-        onErrorAction: (err, exception) => { Exceptions.Add(exception); },
-        disposeTransports: true
-        );
-
-      if (Exceptions.Count != 0)
-      {
-        RaiseNotification($"Failed to send: \n {Exceptions.Last().Message}");
-        return null;
-      }
-
-      if (convertedCount > 0)
-      {
-        var actualCommit = new CommitCreateInput
-        {
-          streamId = streamId,
-          objectId = commitObjId,
-          branchName = state.Branch.name,
-          message = state.CommitMessage != null ? state.CommitMessage : $"Pushed {convertedCount} elements from {Utils.AppName}.",
-          sourceApplication = Utils.VersionedAppName
-        };
-
-        if (state.PreviousCommitId != null) { actualCommit.parents = new List<string>() { state.PreviousCommitId }; }
-
-        try
-        {
-          var commitId = await client.CommitCreate(actualCommit);
-
-          await state.RefreshStream();
-          state.PreviousCommitId = commitId;
-
-          PersistAndUpdateStreamInFile(state);
-          RaiseNotification($"{convertedCount} objects sent to {state.Stream.name}.");
-        }
-        catch (Exception e)
-        {
-          Globals.Notify($"Failed to create commit.\n{e.Message}");
-          state.Errors.Add(e);
-        }
-      }
-      else
-      {
-        Globals.Notify($"Did not create commit: no objects could be converted.");
-      }
-
-      return state;
     }
 
     private List<string> GetObjectsFromFilter(ISelectionFilter filter, ISpeckleConverter converter)
     {
+      var selection = new List<string>();
       switch (filter.Slug)
       {
+        case "manual":
+          return GetSelectedObjects();
         case "all":
           return Doc.ConvertibleObjects(converter);
         case "layer":
-          var layerObjs = new List<string>();
           foreach (var layerName in filter.Selection)
           {
-            AcadDb.TypedValue[] layerType = new AcadDb.TypedValue[1] { new AcadDb.TypedValue((int)AcadDb.DxfCode.LayerName, layerName) };
+            TypedValue[] layerType = new TypedValue[1] { new TypedValue((int)DxfCode.LayerName, layerName) };
             PromptSelectionResult prompt = Doc.Editor.SelectAll(new SelectionFilter(layerType));
             if (prompt.Status == PromptStatus.OK)
-              layerObjs.AddRange(prompt.Value.GetHandles());
+              selection.AddRange(prompt.Value.GetHandles());
           }
-          return layerObjs;
-        default:
-          RaiseNotification("Filter type is not supported in this app. Why did the developer implement it in the first place?");
-          return new List<string>();
+          return selection;
       }
+      return selection;
     }
-
     #endregion
 
     #region events
+    public void RegisterAppEvents()
+    {
+      //// GLOBAL EVENT HANDLERS
+      Application.DocumentWindowCollection.DocumentWindowActivated += Application_WindowActivated;
+      Application.DocumentManager.DocumentActivated += Application_DocumentActivated;
+      Doc.BeginDocumentClose += Application_DocumentClosed;
+    }
+
+    //checks whether to refresh the stream list in case the user changes active view and selects a different document
+    private void Application_WindowActivated(object sender, DocumentWindowActivatedEventArgs e)
+    {
+      if (e.DocumentWindow.Document == null || UpdateSavedStreams == null)
+        return;
+
+      var streams = GetStreamsInFile();
+      UpdateSavedStreams(streams);
+    }
 
     private void Application_DocumentClosed(object sender, DocumentBeginCloseEventArgs e)
     {
@@ -743,22 +713,25 @@ namespace Speckle.ConnectorAutocadCivil.UI
       if (Doc != null)
         return;
 
-      SpeckleAutocadCommand.Bootstrapper.Application.MainWindow.Hide();
-
-      var appEvent = new ApplicationEvent() { Type = ApplicationEvent.EventType.DocumentClosed };
-      NotifyUi(appEvent);
+      if (SpeckleAutocadCommand.MainWindow != null)
+        SpeckleAutocadCommand.MainWindow.Hide();
     }
 
     private void Application_DocumentActivated(object sender, DocumentCollectionEventArgs e)
     {
       // Triggered when a document window is activated. This will happen automatically if a document is newly created or opened.
-      var appEvent = new ApplicationEvent()
-      {
-        Type = ApplicationEvent.EventType.DocumentOpened,
-        DynamicInfo = GetStreamsInFile()
-      };
+      if (e.Document == null)
+        return;
 
-      NotifyUi(appEvent);
+      // transfer old dui streamstates - deprecate this line after a few releases
+      SpeckleStreamManager.TransferOldSpeckleStreams();
+
+      var streams = GetStreamsInFile();
+      if (streams != null && streams.Count != 0)
+      {
+        SpeckleAutocadCommand.CreateOrFocusSpeckle();
+        UpdateSavedStreams(streams);
+      }
     }
     #endregion
   }
